course-03/project/solution/requirements.txt
# UDA-Hub Requirements
# Python 3.12.3
# Agentic Customer Support System with Multi-Agent Architecture

# Core LLM & Agent Framework
langchain==1.2.11
langchain-core==1.2.18
langchain-openai==1.1.11
langgraph==1.1.0
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langchain-community>=0.4.1

# OpenAI Integration
openai==2.26.0

# Vector Store & Embeddings for RAG
chromadb==1.5.5

# Data Processing & Serialization
pydantic==2.12.5
python-dotenv==1.2.2

# Web Scraping & HTTP
beautifulsoup4==4.14.3
requests==2.32.5

# Database ORM
sqlalchemy==2.0.48

# Testing Framework
pytest==9.0.2

# Utilities
click==8.3.1
tqdm==4.67.3
requests-toolbelt==1.0.0

# LangChain Supporting Libraries
langsmith==0.7.16
tenacity==9.1.4

# LLM Utilities
tiktoken==0.12.0

# Core Dependencies (automatically installed by above)
pydantic-core==2.41.5
typing-extensions==4.15.0
httpx==0.28.1
httpcore==1.0.9
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
sniffio==1.3.1
urllib3==2.6.3
